Move most basic functions with minimal dependencies to standalone.js file to ensure googlesitekit-admin.js really only relies on wp-i18n (2/2).
