+++ assets/js/googlesitekit-admin.js
@@ -19,12 +19,12 @@
  */
 
 /**
- * External dependencies
+ * Internal dependencies
  */
 import {
 	appendNotificationsCount,
 	clearAppLocalStorage,
-} from 'GoogleUtil';
+} from './util/standalone';
 
 if ( 'toplevel_page_googlesitekit-dashboard' !== window.pagenow && 'site-kit_page_googlesitekit-splash' !== window.pagenow && 'admin_page_googlesitekit-splash' !== window.pagenow && window.localStorage ) {
 	// The total notifications count should always rely on local storage
